dev: lib: Convert mgebuffer index fields to size_t
Convert the mgebuffer.proc_next and next_free fields to size_t.

+++ dev/src/prg/c/inc/mgebuffer.h
@@ -24,7 +24,9 @@
  * 09/11/2017	MG	1.0.3	Add SPDX license tag.			*
  * 02/01/2018	MG	1.0.4	Move to new source directory structure.	*
  * 04/08/2018	MG	1.0.5	Improve offset field name to proc_next.	*
- *				Improve index field name to next_free.	#
+ *				Improve index field name to next_free.	*
+ *				Convert proc_next and next_free to	*
+ *				size_t.					*
  *									*
  ************************************************************************
  */
@@ -48,10 +50,10 @@ BEGIN_C_DECLS
  * A buffer object.
  */
 struct mgebuffer {
-	char *buffer;	/**< Buffer storage. */
-	size_t size;	/**< Size of the buffer storage area. */
-	int proc_next;	/**< Next buffer location for processing. */
-	int next_free;	/**< Next free buffer location. */
+	char *buffer;		/**< Buffer storage. */
+	size_t size;		/**< Size of the buffer storage area. */
+	size_t proc_next;	/**< Next buffer location for processing. */
+	size_t next_free;	/**< Next free buffer location. */
 };
 
 struct mgebuffer *concat_buf(const char *s_buf, const ssize_t s_buf_os,
